+++ apps/core/src/adapters/cloud/ui.tsx
@@ -11,7 +11,6 @@ import { useCurrentUser } from '../../hooks/affine/use-current-user';
 import { useIsWorkspaceOwner } from '../../hooks/affine/use-is-workspace-owner';
 import { useWorkspace } from '../../hooks/use-workspace';
 import {
-  BlockSuitePageList,
   NewWorkspaceSettingDetail,
   PageDetailEditor,
   Provider,
@@ -59,16 +58,6 @@ export const UI = {
           workspace={workspace.blockSuiteWorkspace}
         />
       </>
-    );
-  },
-  PageList: ({ blockSuiteWorkspace, onOpenPage, collection }) => {
-    return (
-      <BlockSuitePageList
-        listType="all"
-        collection={collection}
-        onOpenPage={onOpenPage}
-        blockSuiteWorkspace={blockSuiteWorkspace}
-      />
     );
   },
   NewSettingsDetail: ({
