+++ apps/core/src/adapters/public-cloud/ui.tsx
@@ -5,7 +5,7 @@ import { initEmptyPage } from '@toeverything/infra/blocksuite';
 import { useCallback } from 'react';
 
 import { useWorkspace } from '../../hooks/use-workspace';
-import { BlockSuitePageList, PageDetailEditor, Provider } from '../shared';
+import { PageDetailEditor, Provider } from '../shared';
 
 export const UI = {
   Provider,
@@ -27,16 +27,6 @@ export const UI = {
           workspace={workspace.blockSuiteWorkspace}
         />
       </>
-    );
-  },
-  PageList: ({ blockSuiteWorkspace, onOpenPage, collection }) => {
-    return (
-      <BlockSuitePageList
-        listType="all"
-        collection={collection}
-        onOpenPage={onOpenPage}
-        blockSuiteWorkspace={blockSuiteWorkspace}
-      />
     );
   },
   NewSettingsDetail: () => {
