+++ apps/core/src/adapters/shared.ts
@@ -14,14 +14,6 @@ export const NewWorkspaceSettingDetail = lazy(() =>
   )
 );
 
-export const BlockSuitePageList = lazy(() =>
-  import('../components/blocksuite/block-suite-page-list').then(
-    ({ BlockSuitePageList }) => ({
-      default: BlockSuitePageList,
-    })
-  )
-);
-
 export const PageDetailEditor = lazy(() =>
   import('../components/page-detail-editor').then(({ PageDetailEditor }) => ({
     default: PageDetailEditor,
