+++ apps/core/src/pages/workspace/all-page.css.ts
@@ -3,5 +3,5 @@ import { style } from '@vanilla-extract/css';
 export const root = style({
   overflow: 'auto',
   height: '100%',
-  paddingBottom: '32px'
+  paddingBottom: '32px',
 });
