+++ apps/core/src/pages/workspace/all-page.tsx
@@ -2,9 +2,8 @@ import { toast } from '@affine/component';
 import {
   CollectionBar,
   OperationCell,
-  useCollectionManager,
+  PageList,
 } from '@affine/component/page-list';
-import { PageList } from '@affine/component/page-list-table';
 import { WorkspaceSubPath } from '@affine/env/workspace';
 import { useAFFiNEI18N } from '@affine/i18n/hooks';
 import { assertExists } from '@blocksuite/global/utils';
@@ -12,23 +11,20 @@ import type { PageMeta } from '@blocksuite/store';
 import { useBlockSuitePageMeta } from '@toeverything/hooks/use-block-suite-page-meta';
 import { getActiveBlockSuiteWorkspaceAtom } from '@toeverything/infra/__internal__/workspace';
 import { getCurrentStore } from '@toeverything/infra/atom';
-import { useAtom } from 'jotai';
-import { useCallback, useMemo } from 'react';
+import { useCallback } from 'react';
 import type { LoaderFunction } from 'react-router-dom';
 import { redirect } from 'react-router-dom';
 
 import { getUIAdapter } from '../../adapters/workspace';
-import { allPageModeSelectAtom } from '../../atoms';
 import { collectionsCRUDAtom } from '../../atoms/collections';
 import { usePageHelper } from '../../components/blocksuite/block-suite-page-list/utils';
 import { useBlockSuiteMetaHelper } from '../../hooks/affine/use-block-suite-meta-helper';
 import { useTrashModalHelper } from '../../hooks/affine/use-trash-modal-helper';
 import { useCurrentWorkspace } from '../../hooks/current/use-current-workspace';
 import { useGetPageInfoById } from '../../hooks/use-get-page-info';
-import { useNavigateHelper } from '../../hooks/use-navigate-helper';
-import { filterPage } from '../../utils/filter';
 import * as styles from './all-page.css';
-import { PageListEmpty } from './page-list-empty';
+import { EmptyPageList } from './page-list-empty';
+import { useFilteredPageMetas } from './pages';
 
 export const loader: LoaderFunction = async args => {
   const rootStore = getCurrentStore();
@@ -49,20 +45,7 @@ export const loader: LoaderFunction = async args => {
 };
 
 export const AllPage = () => {
-  const { jumpToPage } = useNavigateHelper();
   const [currentWorkspace] = useCurrentWorkspace();
-  const { currentCollection } = useCollectionManager(collectionsCRUDAtom);
-  const onClickPage = useCallback(
-    (pageId: string, newTab?: boolean) => {
-      assertExists(currentWorkspace);
-      if (newTab) {
-        window.open(`/workspace/${currentWorkspace?.id}/${pageId}`, '_blank');
-      } else {
-        jumpToPage(currentWorkspace.id, pageId);
-      }
-    },
-    [currentWorkspace, jumpToPage]
-  );
   const { Header } = getUIAdapter(currentWorkspace.flavour);
   const { isPreferredEdgeless } = usePageHelper(
     currentWorkspace.blockSuiteWorkspace
@@ -97,7 +80,7 @@ export const AllPage = () => {
           favorite={!!page.favorite}
           isPublic={!!page.isPublic}
           onDisablePublicSharing={onDisablePublicSharing}
-          onOpenPageInNewTab={() => onClickPage(page.id, true)}
+          link={`/workspace/${currentWorkspace.id}/${page.id}`}
           onRemoveToTrash={() =>
             setTrashModal({
               open: true,
@@ -117,39 +100,15 @@ export const AllPage = () => {
         />
       );
     },
-    [onClickPage, setTrashModal, t, toggleFavorite]
+    [currentWorkspace.id, setTrashModal, t, toggleFavorite]
   );
 
   const getPageInfo = useGetPageInfoById(currentWorkspace.blockSuiteWorkspace);
 
-  const [filterMode] = useAtom(allPageModeSelectAtom);
-
-  const filteredPageMetas = useMemo(
-    () =>
-      pageMetas
-        .filter(pageMeta => {
-          if (filterMode === 'all') {
-            return true;
-          }
-          if (filterMode === 'edgeless') {
-            return isPreferredEdgeless(pageMeta.id);
-          }
-          if (filterMode === 'page') {
-            return !isPreferredEdgeless(pageMeta.id);
-          }
-          console.error('unknown filter mode', pageMeta, filterMode);
-          return true;
-        })
-        .filter(pageMeta => {
-          if (pageMeta.trash) {
-            return false;
-          }
-          if (!currentCollection) {
-            return true;
-          }
-          return filterPage(currentCollection, pageMeta);
-        }),
-    [pageMetas, filterMode, isPreferredEdgeless, currentCollection]
+  const filteredPageMetas = useFilteredPageMetas(
+    'all',
+    pageMetas,
+    currentWorkspace.blockSuiteWorkspace
   );
 
   return (
@@ -171,7 +130,7 @@ export const AllPage = () => {
         pages={filteredPageMetas}
         renderPageAsLink
         fallback={
-          <PageListEmpty
+          <EmptyPageList
             type="all"
             blockSuiteWorkspace={currentWorkspace.blockSuiteWorkspace}
           />
