+++ apps/core/src/pages/workspace/trash-page.tsx
@@ -1,30 +1,60 @@
+import { toast } from '@affine/component';
+import { PageList, TrashOperationCell } from '@affine/component/page-list';
 import { WorkspaceSubPath } from '@affine/env/workspace';
+import { useAFFiNEI18N } from '@affine/i18n/hooks';
 import { assertExists } from '@blocksuite/global/utils';
+import type { PageMeta } from '@blocksuite/store';
+import { useBlockSuitePageMeta } from '@toeverything/hooks/use-block-suite-page-meta';
 import { useCallback } from 'react';
 
 import { getUIAdapter } from '../../adapters/workspace';
-import { BlockSuitePageList } from '../../components/blocksuite/block-suite-page-list';
+import { usePageHelper } from '../../components/blocksuite/block-suite-page-list/utils';
+import { useBlockSuiteMetaHelper } from '../../hooks/affine/use-block-suite-meta-helper';
 import { useCurrentWorkspace } from '../../hooks/current/use-current-workspace';
-import { useNavigateHelper } from '../../hooks/use-navigate-helper';
+import { EmptyPageList } from './page-list-empty';
+import { useFilteredPageMetas } from './pages';
 
 export const TrashPage = () => {
-  const { jumpToPage } = useNavigateHelper();
   const [currentWorkspace] = useCurrentWorkspace();
-  const onClickPage = useCallback(
-    (pageId: string, newTab?: boolean) => {
-      assertExists(currentWorkspace);
-      if (newTab) {
-        window.open(`/workspace/${currentWorkspace?.id}/${pageId}`, '_blank');
-      } else {
-        jumpToPage(currentWorkspace.id, pageId);
-      }
-    },
-    [currentWorkspace, jumpToPage]
-  );
   // todo(himself65): refactor to plugin
   const blockSuiteWorkspace = currentWorkspace.blockSuiteWorkspace;
   assertExists(blockSuiteWorkspace);
   const { Header } = getUIAdapter(currentWorkspace.flavour);
+  const pageMetas = useBlockSuitePageMeta(currentWorkspace.blockSuiteWorkspace);
+  const filteredPageMetas = useFilteredPageMetas(
+    'trash',
+    pageMetas,
+    currentWorkspace.blockSuiteWorkspace
+  );
+  const { restoreFromTrash, permanentlyDeletePage } =
+    useBlockSuiteMetaHelper(blockSuiteWorkspace);
+  const { isPreferredEdgeless } = usePageHelper(
+    currentWorkspace.blockSuiteWorkspace
+  );
+  const t = useAFFiNEI18N();
+  const pageOperationsRenderer = useCallback(
+    (page: PageMeta) => {
+      const onRestorePage = () => {
+        restoreFromTrash(page.id);
+        toast(
+          t['com.affine.toastMessage.restored']({
+            title: page.title || 'Untitled',
+          })
+        );
+      };
+      const onPermanentlyDeletePage = () => {
+        permanentlyDeletePage(page.id);
+        toast(t['com.affine.toastMessage.permanentlyDeleted']());
+      };
+      return (
+        <TrashOperationCell
+          onPermanentlyDeletePage={onPermanentlyDeletePage}
+          onRestorePage={onRestorePage}
+        />
+      );
+    },
+    [permanentlyDeletePage, restoreFromTrash, t]
+  );
   return (
     <>
       <Header
@@ -33,10 +63,18 @@ export const TrashPage = () => {
           subPath: WorkspaceSubPath.TRASH,
         }}
       />
-      <BlockSuitePageList
-        blockSuiteWorkspace={blockSuiteWorkspace}
-        onOpenPage={onClickPage}
-        listType="trash"
+      <PageList
+        pages={filteredPageMetas}
+        renderPageAsLink
+        fallback={
+          <EmptyPageList
+            type="trash"
+            blockSuiteWorkspace={currentWorkspace.blockSuiteWorkspace}
+          />
+        }
+        isPreferredEdgeless={isPreferredEdgeless}
+        blockSuiteWorkspace={currentWorkspace.blockSuiteWorkspace}
+        pageOperationsRenderer={pageOperationsRenderer}
       />
     </>
   );
