--- packages/component/src/components/page-list-table/index.tsx
@@ -1,4 +1,0 @@
-export * from './page-list';
-export * from './page-list-item';
-export * from './page-tags';
-export * from './types';
