+++ packages/component/src/components/page-list/index.tsx
@@ -1,12 +1,14 @@
-export * from './all-page';
 export * from './components/favorite-tag';
 export * from './components/new-page-buttton';
 export * from './components/title-cell';
 export * from './filter';
 export * from './operation-cell';
 export * from './operation-menu-items';
+export * from './page-list';
+export * from './page-list-item';
+export * from './page-tags';
 export * from './styles';
-export * from './type';
+export * from './types';
 export * from './use-collection-manager';
 export * from './utils';
 export * from './view';
